requirements.txt
Django==2.2                                                                            
django-crispy-forms==1.8.1                                                             
django-extensions==2.2.8                                                               
django-markdown-deux==1.0.5                                                            
django-pagedown==2.0.3                                                                 
markdown2==2.3.8                                                                       
Pillow==7.0.0                                                                          
pytz==2019.3                                                                           
six==1.14.0                                                                            
sqlparse==0.3.0